requirements.txt
pymongo==4.11.3
langchain==0.3.25
langchain-aws==0.2.22
langchain-google-genai==2.1.4
langchain-openai==0.3.16
langgraph==0.4.2
langgraph-checkpoint-mongodb==0.1.3
tiktoken==0.9.0
sentence_transformers==4.1.0
tqdm==4.67.1
Pillow==11.1.0
google-generativeai>=0.8.3
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
python-dotenv==1.0.0
email-validator==2.1.0
PyPDF2==3.0.1
deep-translator==1.11.4